+++ modules-services/prov-service-core/src/main/java/org/openprovenance/prov/service/core/config/NoStorageConfiguration.java
@@ -1,7 +1,5 @@
 package org.openprovenance.prov.service.core.config;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -10,30 +8,20 @@ import org.apache.logging.log4j.Logger;
 public class NoStorageConfiguration extends StorageConfiguration {
     private final static Logger logger = LogManager.getLogger(NoStorageConfiguration.class);
 
-    public String dbname = "none";
-    public String index = "none";
-    public String storage = "none";
-    public int cache = -1;
-    public boolean autodelete = true;
-    public int delete_period = 1200;
-
-    public String redis_host = "nohost";
-    public int redis_port = -1;
-    public String mongo_host  = "nohost";
-    public int mongo_port = -1;
-    public String uploaded_filepath="nowhere";
-
     public NoStorageConfiguration() {
         logger.info("NoStorageConfiguration");
+        dbname = "none";
+        index = "none";
+        storage = "none";
+        cache = -1;
+        autodelete = true;
+        delete_period = 1200;
+
+        redis_host = "nohost";
+        redis_port = -1;
+        mongo_host  = "nohost";
+        mongo_port = -1;
+        uploaded_filepath="nowhere";
     }
 
-
-    @Override
-    public String toString() {
-        try {
-            return new ObjectMapper().writeValueAsString(this);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-    }
 }